Refactor the scan commands
