+++ src/commands/scan/cmd-metadata.ts
@@ -1,0 +1,77 @@
+import meow from 'meow'
+import colors from 'yoctocolors-cjs'
+
+import { Spinner } from '@socketsecurity/registry/lib/spinner'
+
+import { getOrgScanMetadata } from './get-full-scan-metadata.ts'
+import { commonFlags, outputFlags } from '../../flags'
+import { AuthError } from '../../utils/errors'
+import { getFlagListOutput } from '../../utils/output-formatting'
+import { getDefaultToken } from '../../utils/sdk'
+
+import type {
+  CliCommandConfig,
+  CliSubcommand
+} from '../../utils/meow-with-subcommands'
+
+const config: CliCommandConfig = {
+  commandName: 'metadata',
+  description: "Get a full scan's metadata",
+  hidden: false,
+  flags: {
+    ...commonFlags,
+    ...outputFlags
+  },
+  help: (parentName, config) => `
+    Usage
+      $ ${parentName} ${config.commandName} <org slug> <scan id>
+
+    Options
+      ${getFlagListOutput(config.flags, 6)}
+
+    Examples
+      $ ${parentName} ${config.commandName} FakeOrg 000aaaa1-0000-0a0a-00a0-00a0000000a0
+  `
+}
+
+export const cmdScanMetadata: CliSubcommand = {
+  description: config.description,
+  hidden: config.hidden,
+  run
+}
+
+async function run(
+  argv: readonly string[],
+  importMeta: ImportMeta,
+  { parentName }: { parentName: string }
+): Promise<void> {
+  const cli = meow(config.help(parentName, config), {
+    argv,
+    description: config.description,
+    importMeta,
+    flags: config.flags
+  })
+
+  const [orgSlug = '', fullScanId = ''] = cli.input
+
+  if (!orgSlug || !fullScanId) {
+    console.error(
+      `${colors.bgRed(colors.white('Input error'))}: Please provide the required fields:\n
+      - Org name as the first argument ${!orgSlug ? colors.red('(missing!)') : colors.green('(ok)')}\n
+      - Full Scan ID to inspect as second argument ${!fullScanId ? colors.red('(missing!)') : colors.green('(ok)')}
+    `
+    )
+    config.help(parentName, config)
+    return
+  }
+
+  const apiToken = getDefaultToken()
+  if (!apiToken) {
+    throw new AuthError(
+      'User must be authenticated to run this command. To log in, run the command `socket login` and enter your API key.'
+    )
+  }
+  const spinnerText = "Getting scan's metadata... \n"
+  const spinner = new Spinner({ text: spinnerText }).start()
+  await getOrgScanMetadata(orgSlug, fullScanId, spinner, apiToken)
+}
